refactor: extract shared rate-limit module across API routes
