+++ src/lib/rate-limit.ts
@@ -1,0 +1,43 @@
+import { NextResponse } from "next/server";
+
+interface Bucket {
+  count: number;
+  resetAt: number;
+}
+
+interface RateLimiterOptions {
+  windowMs: number;
+  max: number;
+}
+
+function getClientIp(req: Request): string {
+  const forwarded = req.headers.get("x-forwarded-for");
+  if (forwarded) {
+    return forwarded.split(",")[0].trim();
+  }
+  return req.headers.get("x-real-ip") ?? "unknown";
+}
+
+export function createRateLimiter({ windowMs, max }: RateLimiterOptions) {
+  const buckets = new Map<string, Bucket>();
+
+  return function checkRateLimit(req: Request): boolean {
+    const ip = getClientIp(req);
+    const now = Date.now();
+    const bucket = buckets.get(ip);
+    if (!bucket || bucket.resetAt < now) {
+      buckets.set(ip, { count: 1, resetAt: now + windowMs });
+      return true;
+    }
+    if (bucket.count >= max) {
+      return false;
+    }
+    bucket.count += 1;
+    return true;
+  };
+}
+
+export const RATE_LIMITED_RESPONSE = NextResponse.json(
+  { error: "rate_limited" },
+  { status: 429 },
+);
